Prevent render preview scroll position reset.
Since switching to relative URLs in the handler, updating the render
preview resulted in a reset of the scroll position. This happened
because the relative url didn’t match the full url in the iframe’s
location property. Fix this by converting the resolving the relative
URL to an absolute URL based on the UI’s location.